Tran-API-1/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "werkzeug>=3.1.4",
]
